Optimizing
Creating a cache for x/y offsets in the buffers and textures to reduce
multiplications in my inner loops. Turns out this is a major
optimization. Never calculate something when you can cache it :)

Also changed my bitmaps back to use a BYTE * instead of DWORD * so
reduce extra conversion steps with rgb multiplications. More minor
adjustments to data flow to follow.