More drawtriangle optimizations and general clean up
Cleaned up all the headers. Also have pushed the scanline triangle
rasterizer to roughly 220fps (260 in release). Not entirely sure there's
much more I can do with it. The 'flat' version without any zbuffer,
lighting or texturing is roughly on par with the halfspace version in
terms of speed. Still haven't figured out how to use the hs rasterizer
in any way other than flat triangles so I can't do a proper comparison.
=\ I'll eventually be moving (parts of at least) the drawtriangle
function to assembler to do a 'final' speedup, but I want to ensure that
the algorithm itself is as tight as possible first.